+++ src/main/java/ddd/caffeine/ratrip/common/jwt/JwtConstants.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.core.jwt;
+package ddd.caffeine.ratrip.common.jwt;
 
 public class JwtConstants {
 	public static final String USER_ID = "USER_ID";
