+++ src/main/java/ddd/caffeine/ratrip/module/auth/application/TokenService.java
@@ -3,9 +3,9 @@ package ddd.caffeine.ratrip.module.auth.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import ddd.caffeine.ratrip.core.jwt.JwtProvider;
-import ddd.caffeine.ratrip.core.jwt.JwtRemover;
-import ddd.caffeine.ratrip.core.jwt.JwtUtil;
+import ddd.caffeine.ratrip.common.jwt.JwtProvider;
+import ddd.caffeine.ratrip.common.jwt.JwtRemover;
+import ddd.caffeine.ratrip.common.jwt.JwtUtil;
 import ddd.caffeine.ratrip.module.auth.application.dto.SignOutDto;
 import ddd.caffeine.ratrip.module.auth.application.dto.TokenReissueDto;
 import ddd.caffeine.ratrip.module.auth.presentation.dto.response.TokenResponseDto;
